Update Delete order page and button (#874)

+++ src/NHSD.GPIT.BuyingCatalogue.WebApp/Areas/Order/Controllers/DeleteOrderController.cs
@@ -26,6 +26,14 @@ namespace NHSD.GPIT.BuyingCatalogue.WebApp.Areas.Order.Controllers
         {
             var order = await orderService.GetOrderThin(callOffId, internalOrgId);
 
+            if (order.OrderStatus == OrderStatus.Completed)
+            {
+                return RedirectToAction(
+                    nameof(OrderController.Summary),
+                    typeof(OrderController).ControllerName(),
+                    new { internalOrgId, callOffId });
+            }
+
             var model = new DeleteOrderModel(internalOrgId, order)
             {
                 BackLink = Url.Action(
@@ -40,6 +48,17 @@ namespace NHSD.GPIT.BuyingCatalogue.WebApp.Areas.Order.Controllers
         [HttpPost]
         public async Task<IActionResult> DeleteOrder(string internalOrgId, CallOffId callOffId, DeleteOrderModel model)
         {
+            if (!ModelState.IsValid)
+                return View(model);
+
+            if (!model.SelectedOption.Value)
+            {
+                return RedirectToAction(
+                    nameof(OrderController.Order),
+                    typeof(OrderController).ControllerName(),
+                    new { internalOrgId, callOffId });
+            }
+
             await orderService.DeleteOrder(callOffId, internalOrgId);
 
             return RedirectToAction(
